Expand upon dataset matching tests (#159)
Add more comprehensive dataset matching test which downloads the dataset and checks for equality. Also added instructions to the README on how to run the test. This should allow people to verify whether dockstring matches the dataset on their system. If it does not match, a comprehensive error message describing the mismatch is given.

---------

Co-authored-by: Austin T
Co-authored-by: mgarort